+++ src/IKVM.Runtime/ConstantFieldWrapper.cs
@@ -36,10 +36,6 @@ using System.Reflection;
 using System.Reflection.Emit;
 #endif
 
-#if IMPORTER
-using IKVM.Tools.Importer;
-#endif
-
 namespace IKVM.Runtime
 {
