+++ src/IKVM.Runtime/DynamicTypeWrapper.FinishedTypeImpl.cs
@@ -32,8 +32,6 @@ using IKVM.Reflection.Emit;
 using IKVM.Tools.Importer;
 
 using Type = IKVM.Reflection.Type;
-using DynamicOrAotTypeWrapper = IKVM.Tools.Importer.AotTypeWrapper;
-using ProtectionDomain = System.Object;
 #else
 using System.Reflection;
 using System.Reflection.Emit;
@@ -49,6 +47,7 @@ namespace IKVM.Runtime
     sealed partial class DynamicTypeWrapper
 #endif
     {
+
         sealed class FinishedTypeImpl : DynamicImpl
         {
