+++ src/IKVM.Runtime/GhostMethodWrapper.cs
@@ -28,8 +28,6 @@ using IKVM.Attributes;
 #if IMPORTER || EXPORTER
 using IKVM.Reflection;
 using IKVM.Reflection.Emit;
-
-using Type = IKVM.Reflection.Type;
 #else
 using System.Reflection;
 using System.Reflection.Emit;
