+++ src/IKVM.Runtime/GhostTag.cs
@@ -27,14 +27,6 @@ using System.Threading;
 
 using IKVM.Attributes;
 
-#if IMPORTER || EXPORTER
-using IKVM.Reflection;
-using IKVM.Reflection.Emit;
-
-using Type = IKVM.Reflection.Type;
-#else
-#endif
-
 namespace IKVM.Runtime
 {
 
@@ -102,6 +94,7 @@ namespace IKVM.Runtime
             throw new global::java.lang.ClassCastException(sb.ToString());
 #endif
         }
+
     }
 
 }
