+++ src/IKVM.Runtime/IKVM.Runtime.csproj
@@ -80,10 +80,6 @@
         <JNITrampolineRuntimeIdentifier Include="osx-arm64" TargetName="aarch64-apple-macosx" Includes="$(JNIHeaderDirShare);$(JNIHeaderDirSolaris);$(JNITrampolineSourceDirectory)include" Arguments="-O0" Condition="$(_EnabledImageRuntimes.Contains(';osx-arm64;'))" />
     </ItemGroup>
 
-    <ItemGroup>
-      <Folder Include="ClassFile\" />
-    </ItemGroup>
-
     <Target Name="CompileJNITrampolines" Inputs="$(JNITrampolineSourceDirectory)ikvm.c;$(JNITrampolineSourceDirectory)ikvm.h" Outputs="$(JNITrampolineIntermediateDirectory)%(JNITrampolineRuntimeIdentifier.Identity).o;$(JNITrampolineIntermediateDirectory)%(JNITrampolineRuntimeIdentifier.Identity).ll" Condition=" '$(TargetFramework)' != '' ">
         <ItemGroup>
             <_JNITrampolinesIncludes Include="%(JNITrampolineRuntimeIdentifier.Includes)" />
