+++ src/IKVM.Runtime/MethodWrapper.cs
@@ -38,10 +38,6 @@ using System.Reflection;
 using System.Reflection.Emit;
 #endif
 
-#if IMPORTER
-using IKVM.Tools.Importer;
-#endif
-
 namespace IKVM.Runtime
 {
