+++ src/IKVM.Runtime/MirandaMethodWrapper.cs
@@ -24,16 +24,9 @@
 using IKVM.Attributes;
 
 #if IMPORTER || EXPORTER
-using IKVM.Reflection;
 using IKVM.Reflection.Emit;
-
-using Type = IKVM.Reflection.Type;
 #else
 using System.Reflection.Emit;
-#endif
-
-#if IMPORTER
-using IKVM.Tools.Importer;
 #endif
 
 namespace IKVM.Runtime
