+++ src/IKVM.Runtime/PrimitiveTypeWrapper.cs
@@ -35,10 +35,6 @@ using System.Reflection;
 using System.Reflection.Emit;
 #endif
 
-#if IMPORTER
-using IKVM.Tools.Importer;
-#endif
-
 namespace IKVM.Runtime
 {
