+++ src/IKVM.Runtime/SimpleFieldWrapper.cs
@@ -26,13 +26,10 @@ using System.Diagnostics;
 #if IMPORTER || EXPORTER
 using IKVM.Reflection;
 using IKVM.Reflection.Emit;
-
-using Type = IKVM.Reflection.Type;
 #else
 using System.Reflection;
 using System.Reflection.Emit;
 using System.Threading;
-
 #endif
 
 #if IMPORTER
