+++ src/IKVM.Runtime/TypeWrapperFactory.cs
@@ -25,13 +25,10 @@
 using System;
 using System.Collections.Generic;
 
-#if NETCOREAPP
-using System.Runtime.Loader;
-#endif
-
 #if IMPORTER || EXPORTER
 using IKVM.Reflection;
 using IKVM.Reflection.Emit;
+
 using Type = IKVM.Reflection.Type;
 using ProtectionDomain = System.Object;
 #else
